Make tty_input return the the amount of data written
Also fix a buffer overflow by checking for room in the buffer before
writing to it.

+++ fs/tty.h
@@ -147,7 +147,7 @@ struct tty {
 };
 
 // if blocking, may return _EINTR, otherwise, may return _EAGAIN
-int tty_input(struct tty *tty, const char *input, size_t len, bool blocking);
+ssize_t tty_input(struct tty *tty, const char *input, size_t len, bool blocking);
 void tty_set_winsize(struct tty *tty, struct winsize_ winsize);
 void tty_hangup(struct tty *tty);
